test: Frame을 컬렉션으로 갖는 BowlingGame 생성
